(TYPE): join struct line by line fix
