+++ src/main/kotlin/org/rust/lang/core/psi/RsPsiFactory.kt
@@ -45,7 +45,6 @@ class RsPsiFactory(private val project: Project) {
         return result
     }
 
-
     fun createBlockExpr(body: String): RsBlockExpr =
         createExpressionOfType("{ $body }")
 
@@ -56,11 +55,6 @@ class RsPsiFactory(private val project: Project) {
         val path = createFromText<RsPathExpr>("fn main() { $text;}")?.path ?: return null
         if (path.text != text) return null
         return path
-    }
-
-    fun createStructLiteralBody(fields: List<PsiElement>): RsStructLiteralBody {
-        val fieldsText = fields.map { e -> e.text!! }.joinToString(", ")
-        return createExpressionOfType<RsStructLiteral>("S { $fieldsText }").structLiteralBody
     }
 
     fun createStructLiteralField(name: String): RsStructLiteralField =
